added like/unlike, spa, prob more

+++ tweets/dislike_tweet.py
@@ -1,10 +1,14 @@
-from bottle import delete, redirect, post, request, response
-import sqlite3
+from bottle import redirect, request, response, delete
+import uuid
 import os
+import jwt
+import time
+import sqlite3
+import imghdr
 
 from settings import get_file_path, check_if_logged_in, time_since_from_epoch, date_text_from_epoch, REGEX_EMAIL, JWT_KEY
 
-@delete("/tweets/delete/<tweet_id>")
+@delete("/tweets/like/<tweet_id>")
 def _(tweet_id):
     try:
         check_if_logged_in()
@@ -15,36 +19,30 @@ def _(tweet_id):
     else:
         db = None
         try:
+            user_id = None
+                # decode jwt cookie to get user id for new tweet
+            user_id = jwt.decode(request.get_cookie("jwt", secret="secret"), JWT_KEY, algorithms=["HS256"])["user_id"]
+            
 
             # connect to database
             db = sqlite3.connect(f"{get_file_path()}/database/database.db")
 
-            # find image and delete it from the folder if it exists
-            tweet_image = db.execute("""
-                SELECT tweet_image
-                FROM tweets
-                WHERE tweet_id = :tweet_id
-                """, (tweet_id,)).fetchone()[0]
-            if tweet_image:
-                if os.path.exists(f"{get_file_path()}/static/images/tweets/{tweet_image}"):
-                    os.remove(f"{get_file_path()}/static/images/tweets/{tweet_image}")
 
             # delete tweet from database
-            db.execute("DELETE FROM tweets WHERE tweet_id = :tweet_id", (tweet_id,))
-
-            # delete tweet likes from database
             db.execute("""
                 DELETE FROM likes
                 WHERE fk_tweet_id = :tweet_id
-                """, (tweet_id,))
-
+                AND fk_user_id = :user_id
+                """, (tweet_id, user_id))
             db.commit()
             return
 
         except Exception as ex:
             print(ex)
             response.status = 500
-            return redirect("/login")
+            return redirect("/home")
+
         finally:
             if db != None:
                 db.close()
+            # return redirect(redirectPath)
